Feature/grpo rewards (#26)
* feat(grpo): add RewardManager with YAML template support and fix entropy bug

- Add RewardManager framework for composable, weighted reward functions
- Support YAML-driven rule templates (rule_template type) for user-editable rewards
- Fix critical bug: compute_entropy() was receiving logits instead of log_probs,
  causing entropy_loss to be ~100M instead of ~0.03
- Add TemplateRuleReward supporting answer_match, tag_check, regex_match modes
- Add RewardModelFunction for external reward model backends
- Add YAML templates for GSM8K format and math rewards
- Add E2E smoke tests (65-67) validating RewardManager equivalence to legacy
- Update grpo_test_1303.md with Run 8 results validating YAML template approach

Run 8 (200 steps) confirms YAML templates produce equivalent training dynamics
to legacy composite_math reward, validating deprecation of legacy reward: path.

* refactor(grpo): consolidate reward system into rewards/ package

- Create ironcore/alignment/rewards/ package with clean structure:
  - base.py: RewardFunction ABC, RewardWorkerPool
  - builtin.py: Math, Code, Format, Keyword, API, Local* functions
  - manager.py: RewardManager with from_config()
  - model.py: RewardModelFunction for API/endpoint backends
  - template.py: TemplateRuleReward for YAML templates
- Remove legacy files: rewards.py, reward_manager.py, reward_model.py, reward_rules.py
- Remove RewardConfig from config_alignment.py, require reward_manager for GRPO
- Merge test_e2e_grpo_smoke.py into test_reward_manager.py (72 tests total)
- Delete 8 legacy config files, keep only FSDP smoke test config
- Add YAML reward templates: format_cot.yaml, format_deepseek.yaml

* fix(grpo): misc fixes and config updates for RewardManager

- Update grpo_gsm8k.yaml to use reward_manager with YAML templates
- KL divergence: tighten clamp from (-10,10) to (-6,6) for stability
- Rotary embedding: ensure sin/cos buffers stay in fp32
- Transformer: fix bias mask tensor device placement
- GRPO trainer: use sum of log-probs (standard policy gradient), remove empty_cache
- Rollout: optimize KV cache expansion with view+expand
- Add bias config to llama.yaml model spec
- Add e2e marker documentation to pyproject.toml
- Update verify_grpo_prereqs.py to use new rewards package

* chore: exclude e2e tests from default CI run

- Add '-m "not e2e"' to pytest addopts in pyproject.toml
- Remove accidentally added excalidraw-diagram-skill submodule

* chore: remove excalidraw diagrams from git tracking

Keep diagrams local only, add to .gitignore

* chore: remove internal AI workflow protocol from public docs

* Refactor and optimize code structure across multiple files

- Removed deprecated `validate_numerical_errors.py` file.
- Enhanced `RolloutBuffer` class in `buffer.py` for better readability and maintainability by formatting conditional statements.
- Improved `GRPODataset` class in `dataset.py` with clearer list comprehensions for padding indices.
- Cleaned up `TransformerModel` class in `transformer.py` by standardizing string quotes and improving comment clarity.
- Streamlined memory cleanup tests in `test_memory_optimization.py` by removing unnecessary blank lines.

* chore: remove obsolete configuration and history files

* Refactor GRPO loss functions and reward management for improved clarity and performance

* Optimize GPU performance for reference log probability accumulation in GRPOTrainer

+++ ironcore/layers/positional_embedding/rotary.py
@@ -57,10 +57,9 @@ class RotaryPositionalEmbedding(nn.Module):
         if max_pos >= self.max_seq_len_cached:
             self._update_rope_cache(int(max_pos) + 1)
 
-        # Index into sin/cos using position_ids: [batch, seq_len, head_dim//2]
-        # and add a dimension for broadcasting across heads
-        sin_emb = self.sin_emb[position_ids].unsqueeze(2).to(x.dtype)
-        cos_emb = self.cos_emb[position_ids].unsqueeze(2).to(x.dtype)
+        # Ensure buffers remain in fp32 even if model.to(lower_precision) is called
+        sin_emb = self.sin_emb.to(torch.float32)[position_ids].unsqueeze(2).to(x.dtype)
+        cos_emb = self.cos_emb.to(torch.float32)[position_ids].unsqueeze(2).to(x.dtype)
 
         x = self.apply_rotary_pos_emb(x, sin_emb, cos_emb)
